chore: Update types export
